Rename pfp_modify cog and group to avatar_mod for clarity
Added in aliases for ease of use too.
