+++ bot/exts/evergreen/avatar_modification/_effects.py
